9_corewar

+++ 9_corewar/libraries/libftprintf/ft_add_num_to_buf.c
@@ -6,7 +6,7 @@
 /*   By: mwuckert <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/15 19:51:02 by mwuckert          #+#    #+#             */
-/*   Updated: 2019/07/21 15:35:08 by mwuckert         ###   ########.fr       */
+/*   Updated: 2020/04/21 21:30:17 by artem            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ static void	ft_add_row_num_to_buf(t_print *sp, char *num)
 	char c;
 
 	c = *CAST_UL(sp->arg) % sp->format.len.base;
-	if (*CAST_UL(sp->arg) > sp->format.len.base - 1)
+	if (*CAST_UL(sp->arg) > (unsigned long)(sp->format.len.base - 1))
 	{
 		*CAST_UL(sp->arg) /= sp->format.len.base;
 		ft_add_row_num_to_buf(sp, num);
